+++ src/BackendGraph.h
@@ -88,7 +88,7 @@ private:
         std::vector<float>& output_val,
         BackendContext& cudnn_context,
         const int tid,
-        const int batch_size = 1
+        const size_t batch_size = 1
     ) override;
 
     std::shared_ptr<conv_descriptor> convolve_init(
@@ -96,7 +96,7 @@ private:
         const int channels,
         const int outputs,
         const int filter_size,
-        const int batch_size = 1
+        const size_t batch_size = 1
     );
 
     std::shared_ptr<conv_descriptor> convolve_no_relu_init(
@@ -104,14 +104,14 @@ private:
         const int channels,
         const int outputs,
         const int filter_size,
-        const int batch_size = 1
+        const size_t batch_size = 1
     );
 
     std::shared_ptr<conv_descriptor> convolve_add_relu_init(
         cudnnHandle_t handle,
         const int channels,
         const int outputs,
-        const int batch_size = 1
+        const size_t batch_size = 1
     );
 
     void push_weights(
